+++ prj/vs2022/CvtOnnx.vcxproj.filters
@@ -276,6 +276,9 @@
     <ClCompile Include="..\..\src\Cvt\OnnxRuntime\ConvertReluNode.cpp">
       <Filter>OnnxRuntime\R</Filter>
     </ClCompile>
+    <ClCompile Include="..\..\src\Cvt\OnnxRuntime\ConvertReshapeNode.cpp">
+      <Filter>OnnxRuntime\R</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\..\src\Cvt\OnnxRuntime\OnnxRuntime.h">
