+++ src/Cvt/OnnxRuntime/Convert.h
@@ -169,6 +169,8 @@ namespace Synet
 
     bool ConvertReluNode(const onnx::NodeProto& node, LayerParam& layer);
 
+    bool ConvertReshapeNode(const onnx::NodeProto& node, bool trans, const LayerParams& layers, const Bytes& original, const OnnxParam& onnxParam, LayerParam& layer, TensorFormatMap* tensorFormatMap);
+
     bool ConvertRoundNode(const onnx::NodeProto& node, LayerParam& layer);
 
     bool ConvertSignNode(const onnx::NodeProto& node, LayerParam& layer);
