+++ src/Cvt/OnnxRuntime/OnnxRuntime.h
@@ -61,59 +61,6 @@ namespace Synet
 
         //-----------------------------------------------------------------------------------------
 
-        bool ConvertReshapeNode(const onnx::NodeProto& node, bool trans, const LayerParams& layers, const Bytes& original, const OnnxParam& onnxParam, LayerParam& layer, TensorFormatMap* tensorFormatMap)
-        {
-            if (!CheckSourceNumber(layer, 2))
-                return false;
-            const LayerParam* src0 = GetLayer(layers, layer.src()[0]);
-            const LayerParam* src1 = GetLayer(layers, layer.src()[1]);
-            if (src0 == NULL || src1 == NULL || src1->type() != LayerTypeMeta)
-                return false;
-            if (src1->meta().type() == MetaTypeStub)
-                src1 = GetLayer(layers, src1->src()[0]);
-            if (src0->type() == Synet::LayerTypeMeta)
-            {
-                layer.type() = Synet::LayerTypeMeta;
-                layer.meta().type() = Synet::MetaTypeReshape;
-            }            
-            else if (src1->meta().type() == MetaTypeConst)
-            {
-                const TensorParam & alpha = src1->meta().alpha();
-                if (alpha.shape().size() != 1)
-                    return false;
-                Shape& shape = layer.reshape().shape();
-                layer.type() = LayerTypeReshape;
-                shape = Shp(alpha.i64().data(), alpha.shape()[0]);
-                layer.src().resize(1);
-                if (trans && CurrentTensorFormat(layers, layer.src(), true, false, true, tensorFormatMap) == TensorFormatNhwc)
-                {
-                    if (shape.size() == 5)
-                    {
-                        shape = Shp( shape[0], shape[3], shape[4], shape[1], shape[2]);
-                    }
-                    if (shape.size() == 4)
-                    {
-                        shape = Shape({ shape[0], shape[2] , shape[3], shape[1] });
-                    }
-                    if (shape.size() == 3)
-                    {
-                        shape = Shape({ shape[0], shape[2] , shape[1] });
-                    }
-                }
-            }
-            else
-            {
-                layer.type() = LayerTypeReshape;
-            }
-            if (onnxParam.setReshapeAxis1() && layer.type() == LayerTypeReshape)
-            {
-                layer.reshape().axis() = 1;
-                //if (layer.reshape().shape().size() > 1 && layer.reshape().shape()[0] == 1)
-                //    layer.reshape().shape().erase(layer.reshape().shape().begin(), layer.reshape().shape().begin() + 1);
-            }
-            return true;
-        }
-
         bool ConvertResizeNode(const onnx::NodeProto& node, const LayerParams& layers, const Bytes& original, LayerParam& layer)
         {
             if (layer.src().size() == 4)
